requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
PyMuPDF
python-dotenv
